+++ Sources/SwiftUIKit/_Deprecated/MenuListItemIcon.swift
@@ -1,0 +1,4 @@
+import SwiftUI
+
+@available(*, deprecated, renamed: "MenuListIcon")
+typealias MenuListItemIcon = MenuListIcon